Relax CMAKE_MANGLE_OBJECT_FILE_NAMES to handle duplicate object names
Some platforms cannot handle duplicate object file names during the
archiving process as the archiver will overwrite previous object
files instead of appending to the archive. This is an undesirable
outcome when specifying all object files to be archived on a single
archive command.

We modify the existing CMAKE_MANGLE_OBJECT_FILE_NAMES option slightly
to mangle the object file names if a duplicate file name exists, in
addition to handling the "+" in the name as was the original intent
of this option.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>